refactor(onboarding): separate components into separate files
style(onboarding): apply designs styles to card
